requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
mediapipe>=0.10.0
facenet-pytorch>=2.5.0
librosa>=0.10.0
scipy>=1.11.0
timm>=0.9.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0
streamlit>=1.28.0
plotly>=5.0.0
tqdm>=4.66.0
pillow>=10.0.0
pyyaml>=6.0.0
albumentations>=1.3.0
